seq: move number parsing to its own module
Create the new `number.rs` module to contain the `Number` enumeration
and the trait implementation for parsing a `Number` from a string.
